fix: respect resourceGroup configuration in IBMNodeClass
The Karpenter IBM Cloud Provider was ignoring the resourceGroup field
specified in IBMNodeClass, causing instances to be created in the
account's default resource group instead of the specified one.

This fix adds the resourceGroup configuration to the instance prototype
when creating VPC instances, ensuring instances are placed in the
correct resource group as specified in the IBMNodeClass.

Fixes #141